Use __dispatch__ selector for catch-all messages
The previous mechanism for looking up the slots associated with a
selector used the `lookup` message, as originally proposed by Piumarta
in his paper.

While elegant, it creates a lot of friction with the design I'm going
for, and the future implementation of traits deriving from multiple
parents.

Instead, if a selector is not found, we search for the special symbol
`__dispatch__`, and provide a default one in Object that panics with
an error.